数据库脚本修改

+++ database/migrations/2017_08_20_114337_create_admins_table.php
@@ -19,7 +19,7 @@ class CreateAdminsTable extends Migration
             $table->string('username', 20)->comment('用户名');
             $table->string('email', 30)->comment('邮箱');
             $table->string('password', 100)->comment('密码');
-            $table->tinyInteger('permission_id')->comment('权限id');
+            $table->integer('permission_id')->comment('权限id');
             $table->text('permission_include')->nullable()->comment('权限节点,空表示未自定义权限');
             $table->char('last_login_ip', 15)->default(0)->comment('最后登录ip');
             $table->timestamp('last_login_time')->nullable()->comment('最后登录时间');
